linking logni and register pages/ Creating VerifiyEmail file
